Adding volunteer award drawing capability. update minutes balance from int to float.

+++ wpa_project/minutes/models/minutes_models.py
@@ -6,11 +6,10 @@ logger = logging.getLogger(__name__)
 
 class Minutes(models.Model):
     meeting_date = models.DateTimeField(default=timezone.now)
-    # start_time = models.DateTimeField(default=timezone.now)
     attending = models.CharField(max_length=250)
     minutes_text = models.CharField(max_length=250)
     memberships = models.IntegerField()
-    balance = models.IntegerField(default=None, null=True)
+    balance = models.FloatField(default=None, null=True)
     discussion = models.TextField()
     end_time = models.DateTimeField(null=True)
